+++ app/src/main/java/com/sun/japaneselisteningtrainer/ui/audio/player/components/Icons.kt
@@ -1,0 +1,60 @@
+package com.sun.japaneselisteningtrainer.ui.audio.player.components
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Edit
+import androidx.compose.material.icons.filled.Favorite
+import androidx.compose.material.icons.filled.FavoriteBorder
+import androidx.compose.material.icons.filled.Shuffle
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+/* --------------------------------- Icons ---------------------------------- */
+
+@Composable
+fun RandomToggleButton(
+    enabled: Boolean,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val cs = MaterialTheme.colorScheme
+    IconButton(onClick = onClick, modifier = modifier) {
+        Icon(
+            Icons.Default.Shuffle,
+            contentDescription = "Shuffle",
+            tint = if (enabled) cs.primary else cs.onSurface
+        )
+    }
+}
+
+@Composable
+fun FavoriteToggleButton(
+    isFavorite: Boolean,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val cs = MaterialTheme.colorScheme
+    IconButton(onClick = onClick, modifier = modifier) {
+        Icon(
+            imageVector = if (isFavorite) Icons.Default.Favorite else Icons.Default.FavoriteBorder,
+            contentDescription = "Favorite",
+            tint = if (isFavorite) cs.error else cs.onSurface
+        )
+    }
+}
+
+@Composable
+fun EditAudioButton(
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    IconButton(onClick = onClick, modifier = modifier) {
+        Icon(
+            Icons.Default.Edit,
+            contentDescription = "Edit audio",
+            tint = MaterialTheme.colorScheme.onSurface
+        )
+    }
+}
