+++ app/src/main/java/com/sun/japaneselisteningtrainer/ui/audio/player/components/PlaybackControl.kt
@@ -1,0 +1,82 @@
+package com.sun.japaneselisteningtrainer.ui.audio.player.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Pause
+import androidx.compose.material.icons.filled.PlayArrow
+import androidx.compose.material.icons.filled.SkipNext
+import androidx.compose.material.icons.filled.SkipPrevious
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+/* --------------------------- Playback controls core ----------------------- */
+
+@Composable
+fun PlaybackControls(
+    isPlaying: Boolean,
+    onPrevious: () -> Unit,
+    onPlayPause: () -> Unit,
+    onNext: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val tint = MaterialTheme.colorScheme.onSurface
+    Row(
+        modifier = modifier,
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.spacedBy(12.dp)
+    ) {
+        IconButton(onClick = onPrevious) {
+            Icon(Icons.Default.SkipPrevious, contentDescription = "Previous", tint = tint)
+        }
+        IconButton(onClick = onPlayPause) {
+            Icon(
+                imageVector = if (isPlaying) Icons.Default.Pause else Icons.Default.PlayArrow,
+                contentDescription = if (isPlaying) "Pause" else "Play",
+                tint = tint,
+                modifier = Modifier.size(44.dp)
+            )
+        }
+        IconButton(onClick = onNext) {
+            Icon(Icons.Default.SkipNext, contentDescription = "Next", tint = tint)
+        }
+    }
+}
+
+@Composable
+fun TransportBar(
+    isPlaying: Boolean,
+    isShuffleOn: Boolean,
+    isFavorite: Boolean,
+    onToggleShuffle: () -> Unit,
+    onPrevious: () -> Unit,
+    onPlayPause: () -> Unit,
+    onNext: () -> Unit,
+    onToggleFavorite: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        modifier = modifier
+            .fillMaxWidth()
+            .padding(horizontal = 4.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.SpaceBetween
+    ) {
+        RandomToggleButton(enabled = isShuffleOn, onClick = onToggleShuffle)
+        PlaybackControls(
+            isPlaying = isPlaying,
+            onPrevious = onPrevious,
+            onPlayPause = onPlayPause,
+            onNext = onNext
+        )
+        FavoriteToggleButton(isFavorite = isFavorite, onClick = onToggleFavorite)
+    }
+}
