+++ app/src/main/java/com/sun/japaneselisteningtrainer/ui/audio/player/components/ProgressBar.kt
@@ -1,0 +1,58 @@
+package com.sun.japaneselisteningtrainer.ui.audio.player.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Slider
+import androidx.compose.material3.SliderDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun AudioProgressBar(
+    progress: Float,                 // 0f..1f
+    currentPosition: Long,
+    duration: Long,
+    onSeek: (Float) -> Unit,
+    modifier: Modifier = Modifier,
+    onSeekFinished: () -> Unit = {}
+) {
+    val currentLabel = formatMs(currentPosition)
+    val durationLabel = formatMs(duration)
+    val cs = MaterialTheme.colorScheme
+    Column(modifier = modifier.fillMaxWidth()) {
+        Slider(
+            value = progress.coerceIn(0f, 1f),
+            onValueChange = onSeek,
+            onValueChangeFinished = onSeekFinished,
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(24.dp),
+            colors = SliderDefaults.colors(
+                thumbColor = cs.primary,
+                activeTrackColor = cs.primary,
+                inactiveTrackColor = cs.onSurface.copy(alpha = 0.24f)
+            )
+        )
+        Row(
+            Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.SpaceBetween
+        ) {
+            Text(currentLabel, color = cs.onSurfaceVariant, fontSize = 12.sp)
+            Text(durationLabel, color = cs.onSurfaceVariant, fontSize = 12.sp)
+        }
+    }
+}
+
+private fun formatMs(ms: Long): String {
+    val sec = (ms / 1000).toInt()
+    val m = sec / 60
+    val s = sec % 60
+    return "%02d:%02d".format(m, s)
+}
