+++ app/src/main/java/com/sun/japaneselisteningtrainer/ui/audio/player/components/RotatingDisc.kt
@@ -1,0 +1,54 @@
+package com.sun.japaneselisteningtrainer.ui.audio.player.components
+
+import androidx.compose.animation.core.LinearEasing
+import androidx.compose.animation.core.RepeatMode
+import androidx.compose.animation.core.animateFloat
+import androidx.compose.animation.core.infiniteRepeatable
+import androidx.compose.animation.core.rememberInfiniteTransition
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun RotatingDisc(
+    imageRes: Int,
+    isPlaying: Boolean,
+    modifier: Modifier = Modifier,
+    size: Dp = 200.dp,
+    borderColor: Color = MaterialTheme.colorScheme.outlineVariant
+) {
+    val infinite = rememberInfiniteTransition(label = "disc")
+    val angle by infinite.animateFloat(
+        initialValue = 0f,
+        targetValue = 360f,
+        animationSpec = infiniteRepeatable(
+            animation = tween(durationMillis = 6000, easing = LinearEasing),
+            repeatMode = RepeatMode.Restart
+        ),
+        label = "angle"
+    )
+
+    Image(
+        painter = painterResource(imageRes),
+        contentDescription = "Album art",
+        contentScale = ContentScale.Crop,
+        modifier = modifier
+            .size(size)
+            .clip(CircleShape)
+            .border(2.dp, borderColor, CircleShape)
+            .rotate(if (isPlaying) angle else 0f)
+    )
+}
