+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,5 @@
 
     <string name="title_activity_test">TestActivity</string>
     <string name="title_activity_main2">MainActivity2</string>
+    <string name="audio_player_title">Player</string>
 </resources>
